refactor: move ManagedAgent up to api/ level alongside ManagedModel
