--- packages/sdk/server-ai/src/api/agent/index.ts
@@ -1,1 +1,0 @@
-export { ManagedAgent } from './ManagedAgent';
