+++ packages/sdk/server-ai/src/api/index.ts
@@ -1,8 +1,8 @@
 export * from './config';
-export * from './agent';
 export * from './chat';
 export * from './graph';
 export * from './judge';
+export * from './ManagedAgent';
 export * from './ManagedModel';
 export * from './metrics';
 export * from './model';
